Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -4,6 +4,6 @@ int main() {
     char u;
     scanf("%c", &u);
 
-    printf("%d", u);
+    printf("%d", (int)u);
     return 0;
 }
